updates for playground
