+++ qmat/playground/diff_eqs/__init__.py
@@ -1,0 +1,10 @@
+
+from qmat.playground.diff_eqs.de_solver import DESolver
+from qmat.playground.diff_eqs.burgers import Burgers
+from qmat.playground.diff_eqs.dahlquist import Dahlquist
+
+__all__ = [
+    "DESolver",
+    "Burgers",
+    "Dahlquist",
+]
