+++ qmat/playground/ex_burgers_integrate.py
@@ -1,0 +1,80 @@
+#
+# Example adopted from `test_4_sdc.py`
+#
+
+
+import numpy as np
+from diff_eqs.burgers import Burgers
+from time_integration.sdc_integration import SDCIntegration
+
+
+from matplotlib import pyplot as plt
+
+N = 128
+nu = 0.2
+
+T: float = 4.0
+domain_size: float = 2.0 * np.pi
+t = np.linspace(0, domain_size, N, endpoint=False)
+
+burgers: Burgers = Burgers(N=N, nu=nu, domain_size=domain_size)
+u0 = burgers.initial_u0("sine")
+
+burgers.run_tests()
+
+
+time_integration = "sdc"
+
+if 1:
+    results = []
+
+    u_analytical = burgers.analytical_integration(u0, t=T)
+
+    for nt in range(4):
+
+        num_timesteps = 2**nt * 1000
+        print(f"Running simulation with num_timesteps={num_timesteps}")
+
+        dt = T / num_timesteps
+
+        burgers: Burgers = Burgers(N=N, nu=nu, domain_size=domain_size)
+        u0 = burgers.initial_u0("sine")
+
+        u = u0.copy()
+
+        if time_integration == "rk1":
+            u = burgers.step_rk1_n(u, dt, num_timesteps)
+
+        elif time_integration == "rk2":
+            u = burgers.step_rk2_n(u, dt, num_timesteps)
+
+        elif time_integration == "rk4":
+            u = burgers.step_rk4_n(u, dt, num_timesteps)
+
+        elif time_integration == "sdc":
+            sdci = SDCIntegration(num_nodes=3, node_type="LEGENDRE", quad_type="LOBATTO")
+            u = sdci.integrate_n(u, dt, num_timesteps, burgers)
+
+        else:
+            raise Exception("TODO")
+
+        plt.plot(t, u, label=f"numerical {num_timesteps}", linestyle="dashed")
+
+        error = np.max(np.abs(u - u_analytical))
+        results.append({"N": num_timesteps, "dt": dt, "error": error})
+
+    prev_error = None
+    for r in results:
+        if prev_error is None:
+            conv = None
+        else:
+            conv = np.log2(prev_error / r["error"])
+
+        print(f"N={r["N"]}, dt={r["dt"]:.6e}, error={r["error"]:.6e}, conv={conv}")
+        prev_error = r["error"]
+
+    if 1:
+        plt.plot(t, u_analytical, label="analytical", linestyle="dotted")
+
+        plt.legend()
+        plt.show()
